Move plugin code from MediaWiki implementation to agnostic interface
The plugin code is itself not related to MediaWiki, so it should live
with the agnostic interface. Especially, since it only mentions that
interface but not any specific implementation.
