+++ tests/integration/NewLexemeForm.test.ts
@@ -4,7 +4,7 @@ import initStore from '@/store';
 import unusedLexemeCreator from '../mocks/unusedLexemeCreator';
 import { ItemSearchKey } from '@/plugins/ItemSearchPlugin/ItemSearch';
 import DevItemSearcher from '@/data-access/DevItemSearcher';
-import { WikiRouterKey } from '@/plugins/WikiRouterPlugin/MediaWikiRouter';
+import { WikiRouterKey } from '@/plugins/WikiRouterPlugin/WikiRouter';
 
 jest.mock( 'lodash/debounce', () => jest.fn( ( fn ) => fn ) );
